+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,33 @@
 
 ### Patch Changes
 
+- ai@6.0.0-beta.46
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [48454ab]
+  - ai@6.0.0-beta.45
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [2b1bf9d]
+  - ai@6.0.0-beta.44
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [27e8c3a]
+  - ai@6.0.0-beta.43
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - ai@6.0.0-beta.42
 
 ## 0.0.1-beta.0